Rename happy number to JAVA3happy number
A happy number is a number defined by the following process:

Starting with any positive integer, replace the number by the sum of the squares of its digits. Repeat the process until the number equals 1 (where it will stay), or it loops endlessly in a cycle which does not include 1.Those numbers for which this process ends in 1 are happy.
Print true if n is a happy number, and false if not.also give me the output of the program
Starting with 19:

1^2 + 9^2 = 1 + 81 = 82
8^2 + 2^2 = 64 + 4 = 68
6^2 + 8^2 = 36 + 64 = 100
1^2 + 0^2 + 0^2 = 1
At this point, the sum becomes 1, and the process stops. Since the process ends with the sum equal to 1, 19 is a happy number.

If the process didn't end with 1 and instead went into a cycle, the number would not be a happy number. In the case of 19, the process converges to 1, making it a happy number.